Move IcsCalendar to it's own file
