+++ com/signals.py
@@ -1,7 +1,8 @@
 from django.db.models.base import post_save
 from django.dispatch import receiver
 
-from com.models import IcsCalendar, News
+from com.calendar import IcsCalendar
+from com.models import News
 
 
 @receiver(post_save, sender=News, dispatch_uid="update_internal_ics")
